Added registration protection and fixed things i forgor

+++ ProjectJon/ADMINISTRATOR.aspx
@@ -4,14 +4,23 @@
 <asp:Content ID="Content2" ContentPlaceHolderID="ContentPlaceHolder1" runat="server">
     <h1>Hello there</h1>
     <form method="post">
-        <pre style="font-family: Impact">Would you like to kill the server</pre>
-        <span>
-                <input type="submit" name="die" value="yes">
+        <%if (Session["admin"] != null)
+            {%>
+            <pre style="font-family: Impact">Would you like to kill the server</pre>
+            <span>
+                    <input type="submit" name="die" value="yes">
 
-        </span>
-        <span>
-                <input type="submit" name="die" value="yes">
+            </span>
+            <span>
+                    <input type="submit" name="die" value="yes">
 
-        </span>
+            </span>
+        <%}
+          else
+          {%>
+            <h1>
+                <marquee class="construction funny" truespeed="" scrolldelay="40" scrollamount="6" style="position: absolute; font-size: 65px">You're not supposed to be here. Be gone</marquee>
+            </h1>
+        <%}%>
     </form>
 </asp:Content>
